mbs: Expose NVR in GraphQL response
Expose the NVR of a module build in the response to the `mbs_build`
query. This makes interaction with the front end slightly easier in that
we don't have to construct the same value from the same components
several times in one request. Moreover, in case we ever need to change,
for instance, how the NVR is constructed, it will only have to be
changed in one place.