Do not remove node_modules from container
Since the application is running on the backed, Node relies on the
`node_modules` directory to be present and populated. It is OK to remove
the directory for front-end applications (e.g. ciboard) where the code
is bundled and served in a couple of static files, but we must keep the
directory here.